pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ConGLUDe"
version = "0.1.0"
description = ""
authors = [
    {name = "Lisa Schneckenreiter", email = "schneckenreiter@ml.jku.at"}
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.9,<3.10"

dependencies = [
    "numpy==1.23.5",
    "scipy==1.10.1",
    "pandas==2.1.4",
    "pytorch-lightning==2.5.3",
    "biopython==1.85",
    "lmdb==1.6.2",
    "einops==0.8.1",
    "statsmodels==0.14.4",
    "hydra-core==1.3.2",
    "hydra-colorlog==1.2.0",
    "transformers==4.51.0",
    "fair-esm==2.0.0",
    "wandb==0.21.1",
    "lifelines==0.30.0",
    "bottleneck==1.4.2",
]


[tool.setuptools]
packages = ["conglude"]
include-package-data = true
